+++ src/paster.ts
@@ -1,12 +1,39 @@
 import * as path from 'path';
 import {mkdir} from 'shelljs';
 import * as clipboard from 'clipboardy'
-import {spawn} from 'child_process';
+import {spawn, ChildProcess} from 'child_process';
 import * as moment from 'moment';
 import * as vscode from 'vscode';
+import {toMarkdown} from './toMarkdown';
 
 enum ClipboardType {
     Unkown = -1, Html = 0, Text, Image
+}
+
+/**
+ * Run command and get stdout
+ * @param shell 
+ * @param options 
+ */
+function runCommand(shell, options: string[]): Promise<string> {
+  return new Promise((resolve, reject) => {
+    let stdout = "";
+    let stderr = "";
+    let process: ChildProcess = spawn(shell, options);
+    process.stdout.on("data", contents => {
+      stdout += contents;
+    });
+    process.stderr.on("data", contents => {
+      stderr += contents;
+    });
+    process.on("error", reject).on("close", function(code) {
+      if (code === 0) {
+        resolve(stdout);
+      } else {
+        reject(new Error(stderr));
+      }
+    });
+  });
 }
 
 class Paster {
@@ -20,7 +47,8 @@ class Paster {
             switch(ctx_type) {
             case ClipboardType.Html:
                 this.pasteTextHtml((html)=>{
-                    var markdown = this.toMarkdown(html);
+                    console.log(html);
+                    var markdown = toMarkdown(html);
                     Paster.writeToEditor(markdown);
                 });
             break;
@@ -158,7 +186,7 @@ class Paster {
         }
 
         if (Paster.isHTML(content)) {
-            return this.toMarkdown(content);
+            return toMarkdown(content);
         }
 
         return content;
@@ -184,142 +212,6 @@ class Paster {
             callback(data);
         });
         return ret;
-    }
-
-    private static toMarkdown(content) {
-        // http://pandoc.org/README.html#pandocs-markdown
-        var pandoc = [
-            {
-                filter: 'h1',
-                replacement: function (content, node) {
-                    var underline = Array(content.length + 1).join('=');
-                    return '\n\n' + content + '\n' + underline + '\n\n';
-                }
-            },
-
-            {
-                filter: 'h2',
-                replacement: function (content, node) {
-                    var underline = Array(content.length + 1).join('-');
-                    return '\n\n' + content + '\n' + underline + '\n\n';
-                }
-            },
-
-            {
-                filter: 'sup',
-                replacement: function (content) {
-                    return '^' + content + '^';
-                }
-            },
-
-            {
-                filter: 'sub',
-                replacement: function (content) {
-                    return '~' + content + '~';
-                }
-            },
-
-            {
-                filter: 'br',
-                replacement: function () {
-                    return '\\\n';
-                }
-            },
-
-            {
-                filter: 'hr',
-                replacement: function () {
-                    return '\n\n* * * * *\n\n';
-                }
-            },
-
-            {
-                filter: ['em', 'i', 'cite', 'var'],
-                replacement: function (content) {
-                    return '*' + content + '*';
-                }
-            },
-
-            {
-                filter: function (node) {
-                    var hasSiblings = node.previousSibling || node.nextSibling;
-                    var isCodeBlock = node.parentNode.nodeName === 'PRE' && !hasSiblings;
-                    var isCodeElem = node.nodeName === 'CODE' ||
-                        node.nodeName === 'KBD' ||
-                        node.nodeName === 'SAMP' ||
-                        node.nodeName === 'TT';
-
-                    return isCodeElem && !isCodeBlock;
-                },
-                replacement: function (content) {
-                    return '`' + content + '`';
-                }
-            },
-
-            {
-                filter: function (node) {
-                    return node.nodeName === 'A' && node.getAttribute('href');
-                },
-                replacement: function (content, node) {
-                    var url = node.getAttribute('href');
-                    var titlePart = node.title ? ' "' + node.title + '"' : '';
-                    if (content === url) {
-                        return '<' + url + '>';
-                    } else if (url === ('mailto:' + content)) {
-                        return '<' + content + '>';
-                    } else {
-                        return '[' + content + '](' + url + titlePart + ')';
-                    }
-                }
-            },
-
-            {
-                filter: 'li',
-                replacement: function (content, node) {
-                    content = content.replace(/^\s+/, '').replace(/\n/gm, '\n    ');
-                    var prefix = '-   ';
-                    var parent = node.parentNode;
-
-                    if (/ol/i.test(parent.nodeName)) {
-                        var index = Array.prototype.indexOf.call(parent.children, node) + 1;
-                        prefix = index + '. ';
-                        while (prefix.length < 4) {
-                            prefix += ' ';
-                        }
-                    }
-
-                    return prefix + content;
-                }
-            },
-            {
-                filter: ['font', 'span', 'div'],
-                replacement: function (content) {
-                    return content;
-                }
-            }
-        ];
-
-        // http://pandoc.org/README.html#smart-punctuation
-        var escape = function (str) {
-            return str.replace(/[\u2018\u2019\u00b4]/g, "'")
-                .replace(/[\u201c\u201d\u2033]/g, '"')
-                .replace(/[\u2212\u2022\u00b7\u25aa]/g, '-')
-                .replace(/[\u2013\u2015]/g, '--')
-                .replace(/\u2014/g, '---')
-                .replace(/\u2026/g, '...')
-                .replace(/[ ]+\n/g, '\n')
-                .replace(/\s*\\\n/g, '\\\n')
-                .replace(/\s*\\\n\s*\\\n/g, '\n\n')
-                .replace(/\s*\\\n\n/g, '\n\n')
-                .replace(/\n-\n/g, '\n')
-                .replace(/\n\n\s*\\\n/g, '\n\n')
-                .replace(/\n\n\n*/g, '\n\n')
-                .replace(/[ ]+$/gm, '')
-                .replace(/^\s+|[\s\\]+$/g, '');
-        };
-
-        var toMarkdown = require("to-markdown");
-        return escape(toMarkdown(content, { converters: pandoc, gfm: true }));
     }
 
     private static pasteImage() {
@@ -505,24 +397,15 @@ class Paster {
             shell = 'sh';
             command = [scriptPath].concat(parameters);
         }
-        const runer = spawn(shell, command);
-        runer.on('exit', function (code, signal) {
-            console.log('exit', code, signal);
-        });
-
-        runer.stdout.on('data', (data: Buffer) => {
+        const runer = runCommand(shell, command);
+        runer.then(stdout => {
             if(callback) {
-                callback(data.toString().trim())
+                callback(stdout.toString().trim());
             }
-        });
-
-        runer.stderr.on('data', (data) => {
-            console.log(shell, command);
-            console.log(`stderr: ${data.toString()}`);
-        });
-
-        runer.on('close', (code) => {
-            console.log(`child process exited with code ${code}`);
+            // return stdout                 // return the command value
+        }, err => {
+            console.log(err);
+            // throw err                     // throw again the error
         });
         return true;
     }